null check titoli meta

+++ curriculum/models.py
@@ -322,8 +322,8 @@ class TitoloPersonale(ModelloSemplice, ConMarcaTemporale, ConAutorizzazioni):
         qualifica_nuova.save()
 
         titolo = kwargs['titolo']
-        print(qualifica_nuova)
-        if titolo.richiede_conferma:
+
+        if titolo.meta and titolo.meta.get('richiede_conferma'):
             sede_attuale = persona.sede_riferimento()
             if not sede_attuale:
                 qualifica_nuova.delete()
